requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
ultralytics>=8.0.0
tkinter
